Optimization page switch by swipe bug 71512
git-svn-id: https://svn.intranet.ydp/ydp/projects/empiria/player/trunk/empiria.player@118768 3e83ccf1-c6ce-0310-9054-8fdb33cf0640

+++ src/eu/ydp/empiria/player/client/controller/multiview/animation/PageSwitchAnimation.java
@@ -15,7 +15,7 @@ public class PageSwitchAnimation extends Animation {
 	private final float animationLength;
 	private double currentPosition = 0;
 
-	public PageSwitchAnimation(FlowPanel element, float to) { // NOPMD
+	public PageSwitchAnimation(FlowPanel element, int to) { // NOPMD
 		this.style = element.getElement().getStyle();
 		this.from = NumberUtils.tryParseFloat(style.getLeft().replaceAll("[a-z%]+$", ""));
 		this.to = to;
